refactor: colocate test files
